requirements.txt
Flask==2.0.1
docxtpl==0.16.7
python-docx==0.8.11